requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
langchain-openai==0.3.3
